+++ common/src/python/centers/center_group.py
@@ -106,12 +106,13 @@ class CenterGroup(CenterAdaptor):
         group = proxy.get_group(group_label=center.name, group_id=center.group)
         assert group, "No group for center"
 
-        center_group = CenterGroup(adcid=center.adcid,
-                                   active=center.active,
-                                   group=group,
-                                   proxy=proxy)
+        center_group = CenterGroup(
+            adcid=center.adcid,
+            active=center.active,  # type: ignore
+            group=group,
+            proxy=proxy)
 
-        tags = list(center.tags)
+        tags = list(center.tags)  # type: ignore
         adcid_tag = f"adcid-{center.adcid}"
         if adcid_tag not in tags:
             tags.append(adcid_tag)