+++ templates/gear/{{cookiecutter.module_name}}/src/python/{{cookiecutter.app_name}}/run.py
@@ -15,7 +15,7 @@ from inputs.parameter_store import ParameterStore
 log = logging.getLogger(__name__)
 
 class {{cookiecutter.class_name}}(GearExecutionEnvironment):
-    """Visitor for the templating gear."""
+    """Visitor for the {{cookiecutter.gear_name}} gear."""
 
     def __init__(self, admin_id: str, client: ClientWrapper, new_only: bool):
         super().__init__(client=client, admin_id=admin_id)
@@ -45,8 +45,8 @@ class {{cookiecutter.class_name}}(GearExecutionEnvironment):
 
     def run(self, context: GearToolkitContext) -> None:
         run(proxy=self.proxy,
-            new_only=.self.__new_only)
-        
+            new_only=self.__new_only)
+
 def main():
     """Main method for {{cookiecutter.gear_name}}."""