+++ docs/pull_metadata/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this gear are documented in this file.
 
 * Adds this CHANGELOG
 * Add error handling for missing config parameters along with typing
+* Updates to use `build_project_map` from `common/src/python/project/project_mapper.py`
 
 ## 0.0.10 and earlier